refactor: delete dead LLMSubsystem, move ObservabilitySubsystem to server/observability/
- Remove LLMSubsystem (dead code: created in factory but never called)
- Remove Subsystem ABC base class (no longer needed)
- Move ObservabilitySubsystem from services/subsystems/ to server/observability/
  where all its consumers live (components.py, pg_metrics_collector.py, _system.py)
- Delete empty services/subsystems/ directory
- Update all imports and tests accordingly
