--- src/nexus/services/subsystem.py
@@ -1,37 +1,0 @@
-"""Subsystem ABC and shared helpers for extracted NexusFS domain services.
-
-Issue #1287: Extract NexusFS Domain Services from God Object.
-
-This module provides:
-- ``Subsystem`` ABC: Base class for all extracted subsystems with lifecycle hooks.
-"""
-
-from abc import ABC, abstractmethod
-from typing import Any
-
-
-class Subsystem(ABC):
-    """Base class for extracted NexusFS subsystems.
-
-    Every subsystem must implement ``health_check()`` returning a dict
-    with at least a ``"status"`` key (``"ok"`` or ``"degraded"``).
-
-    ``cleanup()`` is optional — override if the subsystem holds resources
-    (threads, connections, caches) that need explicit teardown.
-    """
-
-    @abstractmethod
-    def health_check(self) -> dict[str, Any]:
-        """Return health status for this subsystem.
-
-        Returns:
-            Dict with at least ``{"status": "ok"}`` or ``{"status": "degraded", ...}``.
-        """
-        ...
-
-    def cleanup(self) -> None:  # noqa: B027
-        """Release resources held by this subsystem.
-
-        Default implementation is a no-op. Override if you hold threads,
-        connections, or caches that need explicit teardown.
-        """
