--- src/nexus/services/subsystems/__init__.py
@@ -1,12 +1,0 @@
-"""Extracted NexusFS subsystems.
-
-Issue #1287: Extract NexusFS Domain Services from God Object.
-
-Each subsystem is a standalone class extending ``Subsystem`` ABC,
-with explicit constructor dependencies (no self god-reference).
-"""
-
-from nexus.services.subsystems.llm_subsystem import LLMSubsystem
-from nexus.services.subsystems.observability_subsystem import ObservabilitySubsystem
-
-__all__ = ["LLMSubsystem", "ObservabilitySubsystem"]
